+++ examples/example/lib/nested.auto_mappr.dart
@@ -5,9 +5,9 @@
 // AutoMapprGenerator
 // **************************************************************************
 
-// ignore_for_file: type=lint, unnecessary_cast, unused_local_variable
+// ignore_for_file: no_leading_underscores_for_library_prefixes, type=lint
+// ignore_for_file: unnecessary_cast, unused_local_variable
 
-// ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:auto_mappr_annotation/auto_mappr_annotation.dart' as _i1;
 import 'package:examples_example/nested.dart' as _i2;
 
@@ -252,9 +252,8 @@ class $Mappr implements _i1.AutoMapprInterface {
         return null;
       }
       return (_map__i2$NestedTagDto_To__i2$NestedTag(
-            (model as _i2.NestedTagDto?),
-          )
-          as TARGET);
+        (model as _i2.NestedTagDto?),
+      ) as TARGET);
     }
     throw Exception('No ${model.runtimeType} -> $targetTypeOf mapping.');
   }